+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changes:
 -   InputPassword component auto-fill fix
 -   Select component falsy value fix
 -   Table component custom noDataIcon fix, row key fix
+-   TokenLink widget empty image fix
 -   New icons
 
 ## 0.0.33 August 5, 2022